# SwiftShortCourse/document/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.2",
        "title": "iFarm",
        "description": "Restful API List used for communication between Gateway/Mobile and Server",
        "contact": {
            "name": "Techmaster team",
            "email": "cuong@techmaster.vn",
            "url": "https://techmaster.vn"
        }
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "paths": {
...
        "/user": {
            "get": {
                "description": "Get `User` data",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "`User` response\n",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "404": {
                        "description": "`User` not found\n"
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Update `User` data",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "data",
                        "in": "body",
                        "description": "`User` Data\n",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "display_name": {
                                    "type": "string"
                                },
                                "user_phone": {
                                    "type": "string"
                                },
                                "role_ids": {
                                    "type": "array",
                                    "title": "ArrayOfRoleId",
                                    "items": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "user_status": {
                                    "type": "boolean"
                                },
                                "image": {
                                    "type": "string",
                                    "format": "base64"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`User` response\n",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "`User` not found\n"
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "required": [
                "id",
                "user_pass",
                "user_email"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "display_name": {
                    "type": "string"
                },
                "user_pass": {
                    "type": "string",
                    "format": "password"
                },
                "user_email": {
                    "type": "string",
                    "format": "email"
                },
                "user_phone": {
                    "type": "string"
                },
                "user_status": {
                    "type": "boolean"
                },
                "role_ids": {
                    "type": "array",
                    "title": "ArrayOfRoleId",
                    "items": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "user_registered": {
                    "type": "string",
                    "format": "date"
                },
                "user_image_url": {
                    "type": "string"
                },
                "salt": {
                    "type": "string"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}